Rename cekit.type to avoid IntelliJ import clash. Add more test exclusions.
