+++ cekit/builders/buildah.py
@@ -2,8 +2,8 @@ import logging
 import os
 
 from cekit.builder import Builder
+from cekit.cekit_types import DependencyDefinition
 from cekit.tools import run_wrapper
-from cekit.types import DependencyDefinition
 
 LOGGER = logging.getLogger("cekit")
