+++ cekit/builders/podman.py
@@ -3,8 +3,8 @@ import os
 from typing import List
 
 from cekit.builder import Builder
+from cekit.cekit_types import DependencyDefinition
 from cekit.tools import run_wrapper
-from cekit.types import DependencyDefinition
 
 LOGGER = logging.getLogger("cekit")
