+++ cekit/descriptor/module.py
@@ -1,8 +1,8 @@
 from typing import List
 
+from cekit.cekit_types import RawDescriptor
 from cekit.descriptor import Execute, Image
 from cekit.descriptor.image import get_image_schema
-from cekit.types import RawDescriptor
 
 overrides_schema = get_image_schema()
 overrides_schema["map"]["execute"] = {"type": "any"}
