+++ tests/test_integ_install_ordering.py
@@ -1,3 +1,4 @@
+import os
 import platform
 import sys
 
@@ -27,6 +28,9 @@ def run_cekit(image_dir, descriptor, args=None):
 
 @pytest.mark.skipif(
     platform.system() == "Darwin", reason="Disabled on macOS, cannot run Docker"
+)
+@pytest.mark.skipif(
+    os.path.exists("/var/run/docker.sock") is False, reason="No Docker available"
 )
 def test_module_uses_installed_package_in_execute_script(tmp_path, caplog):
     """Check that when a module has an install script that depends on packages that should be installed in an image,
@@ -73,6 +77,9 @@ def test_module_uses_installed_package_in_execute_script(tmp_path, caplog):
 
 @pytest.mark.skipif(
     platform.system() == "Darwin", reason="Disabled on macOS, cannot run Docker"
+)
+@pytest.mark.skipif(
+    os.path.exists("/var/run/docker.sock") is False, reason="No Docker available"
 )
 def test_module_uses_installed_package_in_execute_script_manager_precedence(
     tmp_path, caplog
